Feat: Added a Recent Deaths View For The State of Maryland for Covid19
This view added displays the most recent stats for deaths caused by Covid 19 in the State of Maryland, US. The information is provided by the MD State gov from an open data source. The data is broken down to race, gender, age, and change within the last 24 hours. The Total amount of deaths from Covid 19 is for all counties in the State of MD